+++ lib/primitives/primitives.ml
@@ -6,14 +6,14 @@ open Types
 let wrapprim table =
   (zip (fstl table) (List.map (fun (f, numargs) -> ((fun args _ _ -> f args), numargs)) (sndl table)))
 
-let table: (string * ((evt list -> (evt -> type_wrapper list -> evalopts -> evt) -> evalopts -> evt) * int)) list
+let table: (string * ((evt list -> (evt -> type_wrapper list -> evalstate -> evt) -> evalstate -> evt) * int)) list
   = (wrapprim Numerical.table) @ (wrapprim Dictp.table)
   @ (wrapprim Listp.table) @ (wrapprim Stringp.table) @ Special.table
 
-let unsafe_table: (string * ((evt list -> (evt -> type_wrapper list -> evalopts -> evt) -> evalopts -> evt) * int)) list
+let impure_table: (string * ((evt list -> (evt -> type_wrapper list -> evalstate -> evt) -> evalstate -> evt) * int)) list
   = (wrapprim Iop.table)
 
-let fulltable = unsafe_table @ table
+let fulltable = impure_table @ table
 
 let jsprelude =  Dictp.js ^ Listp.js ^ Special.js