+++ lib/primitives/special.ml
@@ -4,7 +4,7 @@ open Util
 
 (* Special Primitives that are eval-recursive *)
 (** Map a function over an iterable structure *)
-let map args applyfun opts =
+let map args applyfun state =
   let f, s =
     match args with [ f; s ] -> (f, s) | _ -> raise WrongPrimitiveArgs
   in
@@ -12,17 +12,17 @@ let map args applyfun opts =
   match s with
   | EvtList x ->
     EvtList
-      (List.map (fun x -> applyfun f [ AlreadyEvaluated x ] opts) x)
+      (List.map (fun x -> applyfun f [ AlreadyEvaluated x ] state) x)
   | EvtDict d ->
     let keys, values = unzip d in
     EvtDict
       (zip keys
          (List.map
-            (fun x -> applyfun f [ AlreadyEvaluated x ] opts)
+            (fun x -> applyfun f [ AlreadyEvaluated x ] state)
             values))
   | _ -> failwith "Value is not iterable"
 
-let map2 args applyfun opts =
+let map2 args applyfun state =
   let f, s1, s2 =
     match args with
     | [ f; s1; s2 ] -> (f, s1, s2)
@@ -35,11 +35,11 @@ let map2 args applyfun opts =
     EvtList
       (List.map2
          (fun a b ->
-            applyfun f [ AlreadyEvaluated a; AlreadyEvaluated b ] opts)
+            applyfun f [ AlreadyEvaluated a; AlreadyEvaluated b ] state)
          x y)
   | _ -> failwith "Value is not iterable"
 
-let foldl args applyfun opts =
+let foldl args applyfun state =
   let f, a, s =
     match args with
     | [ f; ac; s ] -> (f, ac, s)
@@ -50,17 +50,17 @@ let foldl args applyfun opts =
   | EvtList x ->
     List.fold_left
       (fun acc x ->
-         applyfun f [ AlreadyEvaluated acc; AlreadyEvaluated x ] opts)
+         applyfun f [ AlreadyEvaluated acc; AlreadyEvaluated x ] state)
       a x
   | EvtDict d ->
     let _, values = unzip d in
     List.fold_left
       (fun acc x ->
-         applyfun f [ AlreadyEvaluated acc; AlreadyEvaluated x ] opts)
+         applyfun f [ AlreadyEvaluated acc; AlreadyEvaluated x ] state)
       a values
   | _ -> failwith "Value is not iterable"
 
-let filter args applyfun opts =
+let filter args applyfun state =
   let p, s =
     match args with
     | [ p; s ] -> (p, s)
@@ -72,13 +72,13 @@ let filter args applyfun opts =
     EvtList
       (List.filter
          (fun x ->
-            applyfun p [ AlreadyEvaluated x ] opts = EvtBool true)
+            applyfun p [ AlreadyEvaluated x ] state = EvtBool true)
          x)
   | EvtDict d ->
     EvtDict
       (List.filter
          (fun (_, v) ->
-            applyfun p [ AlreadyEvaluated v ] opts = EvtBool true)
+            applyfun p [ AlreadyEvaluated v ] state = EvtBool true)
          d)
   | _ -> failwith "Value is not iterable"