+++ lib/static.ml
@@ -1,0 +1,2 @@
+(** Contains code for static type checking and various inference functions *)
+